Updated import for logMsg; logMsg was moved to new file logFunctions.js
