+++ backend/reportFunctions.js
@@ -1,5 +1,5 @@
 const echarts = require('echarts');
-const { logMsg } = require('./dbFunctions');
+const { logMsg } = require('./logFunctions');
 const fs = require('fs');
 const path = require('path');
 const pdfMake = require('pdfmake/build/pdfmake');
